refactor: got rid of CheckoutV2
